[Fix] Separate domain layer module
