+++ app/build.gradle
@@ -65,4 +65,5 @@ dependencies {
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.1.5'
 
     implementation project(':datalayer')
+    implementation project(':domainlayer')
 }
