--- app/src/main/java/com/example/cleanarchitecture/domainlayer/requestmodel/EntireDataRequestModel.kt
@@ -1,9 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.requestmodel
-
-import com.example.cleanarchitecture.domainlayer.requestmodel.dataprovidor.CatImageRequestModel
-import com.example.cleanarchitecture.domainlayer.requestmodel.dataprovidor.QuoteRequestModel
-
-data class EntireDataRequestModel(
-    val quoteRequestModel: QuoteRequestModel,
-    val catImageRequestModel: CatImageRequestModel
-)
