--- app/src/main/java/com/example/cleanarchitecture/domainlayer/requestmodel/dataprovidor/QuoteRequestModel.kt
@@ -1,5 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.requestmodel.dataprovidor
-
-import com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.base.GetDataRequestModel
-
-data class QuoteRequestModel(val unit: Unit) : GetDataRequestModel
