--- app/src/main/java/com/example/cleanarchitecture/domainlayer/responsemodel/dataprovidor/CatImageResponseModel.kt
@@ -1,5 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.responsemodel.dataprovidor
-
-import com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.base.GetDataResponseModel
-
-data class CatImageResponseModel(val catImageUrl: String) : GetDataResponseModel
