--- app/src/main/java/com/example/cleanarchitecture/domainlayer/usecase/dataprovidor/base/GetDataInteractor.kt
@@ -1,7 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.base
-
-import io.reactivex.Single
-
-interface GetDataInteractor<RQM : GetDataRequestModel, RPM : GetDataResponseModel> {
-    fun getData(requestModel: RQM): Single<RPM>
-}
