--- app/src/main/java/com/example/cleanarchitecture/domainlayer/usecase/dataprovidor/base/GetDataResponseModel.kt
@@ -1,3 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.base
-
-interface GetDataResponseModel
