--- app/src/main/java/com/example/cleanarchitecture/domainlayer/usecase/dataprovidor/quote/QuoteUseCaseInteractor.kt
@@ -1,7 +1,0 @@
-package com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.quote
-
-import com.example.cleanarchitecture.domainlayer.requestmodel.dataprovidor.QuoteRequestModel
-import com.example.cleanarchitecture.domainlayer.responsemodel.dataprovidor.QuoteResponseModel
-import com.example.cleanarchitecture.domainlayer.usecase.dataprovidor.base.GetDataInteractor
-
-interface QuoteUseCaseInteractor : GetDataInteractor<QuoteRequestModel, QuoteResponseModel>
