+++ domainlayer/src/main/java/com/example/domainlayer/requestmodel/EntireDataRequestModel.kt
@@ -1,0 +1,9 @@
+package com.example.domainlayer.requestmodel
+
+import com.example.domainlayer.requestmodel.dataprovidor.CatImageRequestModel
+import com.example.domainlayer.requestmodel.dataprovidor.QuoteRequestModel
+
+data class EntireDataRequestModel(
+    val quoteRequestModel: QuoteRequestModel,
+    val catImageRequestModel: CatImageRequestModel
+)
