+++ domainlayer/src/main/java/com/example/domainlayer/requestmodel/dataprovidor/CatImageRequestModel.kt
@@ -1,0 +1,6 @@
+package com.example.domainlayer.requestmodel.dataprovidor
+
+import com.example.domainlayer.usecase.dataprovidor.base.GetDataRequestModel
+
+
+data class CatImageRequestModel(val unit: Unit) : GetDataRequestModel
