+++ domainlayer/src/main/java/com/example/domainlayer/responsemodel/dataprovidor/CatImageResponseModel.kt
@@ -1,0 +1,6 @@
+package com.example.domainlayer.responsemodel.dataprovidor
+
+import com.example.domainlayer.usecase.dataprovidor.base.GetDataResponseModel
+
+
+data class CatImageResponseModel(val catImageUrl: String) : GetDataResponseModel
