+++ domainlayer/src/main/java/com/example/domainlayer/responsemodel/dataprovidor/QuoteResponseModel.kt
@@ -1,0 +1,6 @@
+package com.example.domainlayer.responsemodel.dataprovidor
+
+import com.example.domainlayer.usecase.dataprovidor.base.GetDataResponseModel
+
+
+data class QuoteResponseModel(val quote: String) : GetDataResponseModel
