+++ domainlayer/src/main/java/com/example/domainlayer/usecase/dataprovidor/base/GetDataInteractor.kt
@@ -1,0 +1,7 @@
+package com.example.domainlayer.usecase.dataprovidor.base
+
+import io.reactivex.Single
+
+interface GetDataInteractor<RQM : GetDataRequestModel, RPM : GetDataResponseModel> {
+    fun getData(requestModel: RQM): Single<RPM>
+}
