+++ domainlayer/src/main/java/com/example/domainlayer/usecase/dataprovidor/base/GetDataMapper.kt
@@ -1,0 +1,7 @@
+package com.example.domainlayer.usecase.dataprovidor.base
+
+import com.example.datalayer.base.Entity
+
+interface GetDataMapper<E: Entity, RPM: GetDataResponseModel> {
+    fun entityToResponseModel(entity: E): RPM
+}
